Add files via upload
Fixed two row numbering issues where the daily % change loop wasn't taking into account the header row and initial row so the calculation was 2 rows short.  I also fixed an issue where the beta and correlation calculation were also not taking these first two rows into account and were calculating 2 rows short.  These were coming up short because I use the Calculation Period setting as the end of the range, but the last row is actually the Calculation period plus 2.

Fixed an issue where if the time period was changed new data wouldn't be calculated if the most recent timestamp in the data was for today.  Changing the time period will now automatically clear the data for all symbols and cause all data to be recalculated.

Fixed an issue where if a symbol had no data or all data was 0 the correlation function would cause a crash.  This error is now caught and correlation generates a 0 if there is no data.